finance-backend/requirements.txt
google-genai
fastapi
uvicorn
pydantic
pandas
numpy
python-dotenv
scikit-learn
joblib